Issue/34 (#35)

+++ inst/scripts/dump_release.R
@@ -67,7 +67,10 @@ readFiles <- function(links, delay = 60)
     exps <- subset(exps, State == "Complete")
     sigs <- readr::read_csv("sig.csv")
     sigs <- subset(sigs, State == "Complete") 
-    file.remove(c("sig.csv", "exp.csv", "stud.csv"))
+    # If not GitHub Action with BUGSIGDB_TIMESTAMP
+    if (Sys.getenv("BUGSIGDB_TIMESTAMP") != "") {
+        file.remove(c("sig.csv", "exp.csv", "stud.csv"))
+    }
     print(gettextf("Successfully read csv files"))
 
     ind <- setdiff(colnames(studs), c("Reviewer", "State"))
